working on passing input to tool
