+++ codegen/agent.py
@@ -1,0 +1,40 @@
+from langchain.agents import Tool, AgentExecutor, BaseSingleActionAgent
+from langchain import OpenAI
+from typing import List, Tuple, Any, Union
+from langchain.schema import AgentAction, AgentFinish
+
+class CodegenAgent(BaseSingleActionAgent):
+    @property
+    def input_keys(self):
+        return ["input"]
+    
+    def plan(
+        self, intermediate_steps: List[Tuple[AgentAction, str]], **kwargs: Any
+    ) -> Union[AgentAction, AgentFinish]:
+        """Given input, decide what to do.
+
+        Args:
+            intermediate_steps: Steps the LLM has taken to date,
+                along with observations
+            **kwargs: User inputs.
+
+        Returns:
+            Action specifying what tool to use.
+        """
+        print(self.input_keys)
+        return AgentAction(tool="Get Directory Structure", tool_input="any", log="")
+
+    async def aplan(
+        self, intermediate_steps: List[Tuple[AgentAction, str]], **kwargs: Any
+    ) -> Union[AgentAction, AgentFinish]:
+        """Given input, decide what to do.
+
+        Args:
+            intermediate_steps: Steps the LLM has taken to date,
+                along with observations
+            **kwargs: User inputs.
+
+        Returns:
+            Action specifying what tool to use.
+        """
+        return AgentAction(tool="Search", tool_input="foo", log="")
