+++ codegen/codegen.py
@@ -1,49 +1,5 @@
-from langchain.agents import Tool, AgentExecutor, BaseSingleActionAgent
-from langchain import OpenAI
+from .tools.filesystem import GetDirectoryStructure
+from agent import CodegenAgent
 
-tools = [
-    Tool(
-        name = "Search",
-        func=search.run,
-        description="useful for when you need to answer questions about current events",
-        return_direct=True
-    )
-]
-
-from typing import List, Tuple, Any, Union
-from langchain.schema import AgentAction, AgentFinish
-
-class CodegenAgent(BaseSingleActionAgent):
-    @property
-    def input_keys(self):
-        return ["input"]
-    
-    def plan(
-        self, intermediate_steps: List[Tuple[AgentAction, str]], **kwargs: Any
-    ) -> Union[AgentAction, AgentFinish]:
-        """Given input, decide what to do.
-
-        Args:
-            intermediate_steps: Steps the LLM has taken to date,
-                along with observations
-            **kwargs: User inputs.
-
-        Returns:
-            Action specifying what tool to use.
-        """
-        return AgentAction(tool="Search", tool_input="foo", log="")
-
-    async def aplan(
-        self, intermediate_steps: List[Tuple[AgentAction, str]], **kwargs: Any
-    ) -> Union[AgentAction, AgentFinish]:
-        """Given input, decide what to do.
-
-        Args:
-            intermediate_steps: Steps the LLM has taken to date,
-                along with observations
-            **kwargs: User inputs.
-
-        Returns:
-            Action specifying what tool to use.
-        """
-        return AgentAction(tool="Search", tool_input="foo", log="")
+tools = [GetDirectoryStructure]
+agent = CodegenAgent()
